Continued Fleshing Out The Chore View/Controller
Still working on finishing up the chores part of this project. Still
need to work though finishing up the edits and testing the delete. Also
want to get the seeder filled out so I can have some more data to test
with. Added a modal component but still need to further break it down
and also could consolidate more of our buttons.