Major Style Update & Charts & Tabs & Fixed Edits
Wanted to fix the chore edits so we would deal with html encoded chars.
Also updated the style on most pages so that we have a more uniform
color and style. Also added tabs to the dash board but still need to
flesh themout. Added charts to the history page but still need to flesh
them out as well.